refactor(hooks): rename useTaskActions to camelCase and update imports
