+++ src/components/task-board/task-sheet/MarkCompleteButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/components/ui/button";
-import { useTaskActions } from "@/hooks/use-task-actions";
+import { useTaskActions } from "@/hooks/useTaskActions";
 import { cn } from "@/lib/utils";
 import { useTaskStore } from "@/store/useTaskStore";
 import { CheckCircle } from "lucide-react";
